refactor: Exception 위치 변경
- NotFoundException은 application에서 사용하는 예외로 application으로 위치 변경
- UnauthorizedException은 web 요청에서 권한이 없을 때 발생하는 예외로 adaptor로 위치 변경
