+++ src/main/java/com/srltas/runtogether/adapter/in/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.srltas.runtogether.application.exception;
+package com.srltas.runtogether.adapter.in.exception;
 
 import com.srltas.runtogether.common.exception.RunTogetherException;
 import com.srltas.runtogether.common.exception.code.ErrorCode;
