+++ src/main/java/com/srltas/runtogether/application/exception/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package com.srltas.runtogether.adapter.out.exception;
+package com.srltas.runtogether.application.exception;
 
 import com.srltas.runtogether.common.exception.RunTogetherException;
 import com.srltas.runtogether.common.exception.code.ErrorCode;
